proteins and mrnas modules
